+ Added the first analyzer, an alanlyzer which will parse each message and pull out the relative times and dates in order to convert them to UTC. This will require more work, but the framework for it is in + Analyzers will now actually be run and are included into the framework for Phil + Phil will now always ignore bots when bots send messages + Analyzers now return a KVP object like commands do, rather than an array, so that we have knowledge of their names
NOTE: In order to finish our code for timezone-converter, the single most important piece is that we need to know which timezone the user is in. This will either need to be something we get from discord if possible, or something that Phil needs to be manually set with, like he is with birthdays.