+ Added "features"
Previously, we had the ability to enable/disable the daily prompt from being posted or continuing forward. However, that was very specific to the prompts and was not at all extensible. We wanted to do the same thing for timezone processing, so everything was rewritten into features: a generic structure where we can say "are we using XXXX" feature, specify just a little bit about that feature (basically just a name and a database key), and then we've got a new hook set up and commands that can process it, which will make turning on and off some of these features VERY easy.

+ p!enableprompts was renamed to p!enable, and reworked to work with features
+ p!disableprompts was renamed to p!disable, and reworked to work with features
+ Everything previously specific to using the prompts-specific code for enabling/disabling was now reworked to use ./phil/features.js
+ Took some code that was used for the p!request function to stitch together an array and moved it to bot_utils in order to share it with other locations.