+ Timezone questionnaire has undergone MAJOR revisions. Now, instead of having the user provide their UTC offset and whether they observe daylight savings or not, the user will first provide the name of their country, and then choose from a list of timezones in their country if their country has multiple timezones. If their country only has one timezone, it'll automatically select it.
NOTE: Country stage is finished, but Specification and Confirmation stages still need to be implemented.

NOTE: data/ is generated via scripts/. However, it's static data and there isn't a reason why we wouldn't just submit it to the repository directly. The node modules added to package.json are only used within the new script, and their output is stored in a Phil-specific way inside of data/country-timezones.json