Moved the following fields from process.env to ServerConfigs: + ADMIN_CHANNEL_ID + INTRODUCTIONS_CHANNEL_ID + NEWS_CHANNEL_ID + ADMIN_ROLE_ID
If a channel is not provided/doesn't exist, it will now fall through to what Discord considers to be the system channel (which can change). Should that not exist, it'll fall through to the first channel in the server.

Updated the print statement displayed every time that a user uses a command