Pronouns were moved to being a ServerConfig thing. Previously, they were hardcoded inside of BotUtils, which not only made adding more pronouns difficult, but it wouldn't work for multiple servers because they were tightly coupled with the process environment configuration. This is a more dynamic way that appears to check out but which we'll be testing in the next change when we add a new p!pronoun command to change your pronoun.
One of the changes that comes from this change is that you can only have one pronoun specified. This is because ULTIMATELY i'd like to turn it over and have the pronouns be added on a server-by-server basis (have some constant ones and then allow for server-custom ones as well) and trying to figure out which ones take dominance over another if a user has multiple ones sounds SUPER tough to do.