Refactored requestables and how you create them to now be much cleaner. Each requestable now has its own factory which is responsible for determining what the handle is and what emoji should be displayed. Rather than forcing the caller to create a reactable and then also add the reactions manually, the factory will do all of that setup work!
More work will be coming in terms of cleaning up previous requestables. As a general rule of thumb, I don't see a reason that a user would ever need to have more than one requestable of a particular type active at a time.

Reactables no longer require a server in order to work (there was no need for it).