Began exploring ValueInterpreters
This is a new, safer, more comprehensive and reusable way to handle user input. This will also be used to power the p!config and p!bucket commands that will be coming online soon. I might have to remove the generic type parameter, though, in order to allow for using these in other non-homogeneous arrays/types.