Added a p!config (for server config) and a configuration command base.
The base is built to be reusable and for display and mutation, which will be usable later for a rework of p!bucket. The general flow is all down and the properties are able to work, which means that next steps are: a) adding all of the pretty output and b) adding in the properties to p!config and c) reworking p!bucket to use this.