GlobalConfig is now a global singleton
I'm not super happy with this change, but the more and more that we were needing to pipe around this universal instance of something that was based off of the immutable global environment, the more that a change like this was going to be come necessary. The argument against it is it hides away dependencies and it'd make it harder to mock, but we have TypeScript to still show us where this is being used (and it's used in few places) and we're not DOING unit testing right now, and we'd be providing an artificial environment for it when/if we get there anyways.