version 14: merge in completed (#1)
* [Version 14]

* Began exploring ValueInterpreters

This is a new, safer, more comprehensive and reusable way to handle user input. This will also be used to power the p!config and p!bucket commands that will be coming online soon. I might have to remove the generic type parameter, though, in order to allow for using these in other non-homogeneous arrays/types.

* Added a p!config (for server config) and a configuration command base.

The base is built to be reusable and for display and mutation, which will be usable later for a rework of p!bucket. The general flow is all down and the properties are able to work, which means that next steps are: a) adding all of the pretty output and b) adding in the properties to p!config and c) reworking p!bucket to use this.

* Changed ValueInterpreters -> TypeDefinitions in order to allow them to store more information about their particular types.

ConfigCommandBase now is able to display all of the current values and the various properties as well as descriptions about them.

p!config has more properties in it.

Added proper typing to DiscordPromises.sendEmbedMessage

* GlobalConfig is now a global singleton

I'm not super happy with this change, but the more and more that we were needing to pipe around this universal instance of something that was based off of the immutable global environment, the more that a change like this was going to be come necessary. The argument against it is it hides away dependencies and it'd make it harder to mock, but we have TypeScript to still show us where this is being used (and it's used in few places) and we're not DOING unit testing right now, and we'd be providing an artificial environment for it when/if we get there anyways.

* p!config can now change command prefix

Also, lots of changes to ServerConfig following the previous refactors that couldn't be committed until now.

* + Added detailed explanation for how to use a config command
+ 'InteractionType' is now 'Action'
+ Added a 'Display' action that will show all of the property values (previously was the response for not providing an action, now replaced with a detailed explanation of how to use p!config or p!bucket)

* + Implemented the 'info' action on the p!config command, which displays the description of the property, the current value, the default value, the rules for valid values, and examples of how to set/modify this property.
+ Introduced a minimum character length for command prefixes
+ Command prefixes can no longer have any form of whitespace in them.
- Updated the styling for when we format a null channel to display format

* + p!config action 'info' is now canonically 'explain'
+ Clarified the descriptions for the p!config properties that already exist

* + Embeds now use an enum for colour instead of an arbitrary colour, in order to ease development and encourage consistency/reuse
+ Defined the error response for the p!config command for invalid input.
+ The Actions enum in for the configuration command base is now backed by the appropriate values for each action rather than arbitrary numbers.

* Added error explanation for unrecognized action.

Submitting early because I'm going to do a large refactor to clean this file up and hopefully reduce its complexity and bloated scope.

* Broke up the monolithic ConfigCommandBase, separating out each individual action into a separate file, with the full list of actions supported for the ConfigCommandBase being passed in through the constructor. In addition to allowing for better structuring of code, this is what will allow for dynamically adding new actions to the p!bucket command that aren't available in p!config.

* Updated the generation of example invocations for each action (that is, the string that displays a valid example usage of each action for p!config) to now handle arbitrary parameters correctly. This will be pivotal later as new actions require new parameters for other commands.

* Post 'master' merge updating the code to work alongside version-14's changes.

* + p!config, when not specifying a property, will now display a list of all of the properties available to configure as well as examples of using them with the already-specified action
+ Fixed the "cannot use command" message to now display the correct required-but-not-met permission level.

* + Filled out the last TODO message for the ConfigCommandBase - the success message for setting/resetting a property
+ Attempted to clarify the 'display' config action

* + Added the ability to change the admin role to p!config

* added member and welcome-message type definitions

* added p!welcome that allows you to preview a welcome message

* configure welcome message

* start work on reusable prompt transition

* switch to postgrator, yarn, and fix up eslint/prettier

* finish initial database conversion using separate script (bleh)

* begin working on separating the submissions and prompts in code

* initial pass at updating all code to work with new prompt_v2/submission

* have phil post the dustiest submission if no prompts in the queue

* update various locations to point to new prompt/suggestion tables

* fix missing bold close tag in submission session wrapup

* extend duration of submission session from 10min to 25min

* allow UserTimezone.getForUser to return null if user hasn't started

* Allow p!utc with no time provided (defaults to "now")

* clear direct message processes before starting a new one

* consolidate enable/disable code, display feature list if not provided